requirements.txt
requests
pyyaml
websocket-client
slackclient